clearnetwork/pyproject.toml
[project]
name = "clearnetwork"
version = "0.1.0"
description = "Healthcare Insurance Network Intelligence Platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.6.0",
    "geoalchemy2>=0.15.0",
    "python-dotenv>=1.0.0",
    "ijson>=3.2.0",
    "aiohttp>=3.9.0",
    "aiofiles>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
